Cleans up compat.h, moves CrewClass and PCX code to common and ShapeClass header to common.
