+++ common/filepcx.h
@@ -71,7 +71,6 @@ typedef struct
 class GraphicBufferClass;
 
 GraphicBufferClass* Read_PCX_File(char* name, char* palette = NULL, void* buff = NULL, long size = 0);
-GraphicBufferClass* Read_PCX_File(char* name, BufferClass& Buff, char* palette = NULL);
 int Write_PCX_File(char* name, GraphicViewPortClass& pic, unsigned char* palette);
 
 #endif
