+++ redalert/winstub.cpp
@@ -727,7 +727,7 @@ void Memory_Error_Handler(void)
     exit(1);
 }
 
-GraphicBufferClass* Read_PCX_File(char* name, char* Palette, void* Buff, long Size);
+#include "filepcx.h"
 void Load_Title_Screen(char* name, GraphicViewPortClass* video_page, unsigned char* palette)
 {
 
@@ -739,141 +739,4 @@ void Load_Title_Screen(char* name, GraphicViewPortClass* video_page, unsigned ch
         load_buffer->Blit(*video_page);
         delete load_buffer;
     }
-}
-
-#include "common/filepcx.h"
-
-/***************************************************************************
- * READ_PCX_FILE -- read a pcx file into a Graphic Buffer                  *
- *                                                                         *
- *	GraphicBufferClass* Read_PCX_File (char* name, char* palette ,void *Buff, long size );	*
- *  																								*
- *                                                                         *
- * INPUT: name is a NULL terminated string of the format [xxxx.pcx]        *
- *        palette is optional, if palette != NULL the the color palette of *
- *					 the pcx file will be place in the memory block pointed	   *
- *               by palette.																*
- *			 Buff is optional, if Buff == NULL a new memory Buffer		 		*
- *					 will be allocated, otherwise the file will be placed 		*
- *					 at location pointed by Buffer;										*
- *			Size is the size in bytes of the memory block pointed by Buff		*
- *				  is also optional;															* * OUTPUT: on success a
- *pointer to a GraphicBufferClass containing the     * pcx file, NULL otherwise.                                       *
- *																									*
- * WARNINGS:                                                               *
- *         Appears to be a comment-free zone                               *
- *                                                                         *
- * HISTORY:                                                                *
- *   05/03/1995 JRJ : Created.                                             *
- *   04/30/1996 ST : Tidied up and modified to use CCFileClass             *
- *=========================================================================*/
-
-#define POOL_SIZE 2048
-#define READ_CHAR()                                                                                                    \
-    *file_ptr++;                                                                                                       \
-    if (file_ptr >= &pool[POOL_SIZE]) {                                                                                \
-        file_handle.Read(pool, POOL_SIZE);                                                                             \
-        file_ptr = pool;                                                                                               \
-    }
-
-GraphicBufferClass* Read_PCX_File(char* name, char* palette, void* Buff, long Size)
-{
-    unsigned i, j;
-    unsigned rle;
-    unsigned color;
-    unsigned scan_pos;
-    char* file_ptr;
-    int width;
-    int height;
-    char* buffer;
-    PCX_HEADER header;
-    RGB* pal;
-    char pool[POOL_SIZE];
-    GraphicBufferClass* pic;
-
-    CCFileClass file_handle(name);
-
-    if (!file_handle.Is_Available())
-        return (NULL);
-
-    file_handle.Open(READ);
-
-    file_handle.Read(&header, sizeof(PCX_HEADER));
-
-    if (header.id != 10 && header.version != 5 && header.pixelsize != 8)
-        return NULL;
-
-    width = header.width - header.x + 1;
-    height = header.height - header.y + 1;
-
-    if (Buff) {
-        buffer = (char*)Buff;
-        i = Size / width;
-        height = MIN((int)i - 1, height);
-        pic = new GraphicBufferClass(width, height, buffer, Size);
-        if (!(pic && pic->Get_Buffer()))
-            return NULL;
-    } else {
-        pic = new GraphicBufferClass(width, height, NULL, width * (height + 4));
-        if (!(pic && pic->Get_Buffer()))
-            return NULL;
-    }
-
-    buffer = (char*)pic->Get_Buffer();
-    file_ptr = pool;
-    file_handle.Read(pool, POOL_SIZE);
-
-    if (header.byte_per_line != width) {
-
-        for (scan_pos = j = 0; j < (unsigned)height; j++, scan_pos += width) {
-            for (i = 0; i < (unsigned)width;) {
-                rle = READ_CHAR();
-                if (rle > 192) {
-                    rle -= 192;
-                    color = READ_CHAR();
-                    ;
-                    memset(buffer + scan_pos + i, color, rle);
-                    i += rle;
-                } else {
-                    *(buffer + scan_pos + i++) = (char)rle;
-                }
-            }
-        }
-
-        if (i == width)
-            rle = READ_CHAR();
-        if (rle > 192)
-            rle = READ_CHAR();
-
-    } else {
-
-        for (i = 0; i < (unsigned)width * height;) {
-            rle = READ_CHAR();
-            rle &= 0xff;
-            if (rle > 192) {
-                rle -= 192;
-                color = READ_CHAR();
-                memset(buffer + i, color, rle);
-                i += rle;
-            } else {
-                *(buffer + i++) = (char)rle;
-            }
-        }
-    }
-
-    if (palette) {
-        file_handle.Seek(-(256 * (int)sizeof(RGB)), SEEK_END);
-        file_handle.Read(palette, 256L * sizeof(RGB));
-
-        pal = (RGB*)palette;
-        for (i = 0; i < 256; i++) {
-            pal->red >>= 2;
-            pal->green >>= 2;
-            pal->blue >>= 2;
-            pal++;
-        }
-    }
-
-    file_handle.Close();
-    return pic;
 }
