--- tiberiandawn/crew.h
@@ -1,71 +1,0 @@
-//
-// Copyright 2020 Electronic Arts Inc.
-//
-// TiberianDawn.DLL and RedAlert.dll and corresponding source code is free
-// software: you can redistribute it and/or modify it under the terms of
-// the GNU General Public License as published by the Free Software Foundation,
-// either version 3 of the License, or (at your option) any later version.
-
-// TiberianDawn.DLL and RedAlert.dll and corresponding source code is distributed
-// in the hope that it will be useful, but with permitted additional restrictions
-// under Section 7 of the GPL. See the GNU General Public License in LICENSE.TXT
-// distributed with this program. You should have received a copy of the
-// GNU General Public License along with permitted additional restrictions
-// with this program. If not, see https://github.com/electronicarts/CnC_Remastered_Collection
-
-/* $Header:   F:\projects\c&c\vcs\code\crew.h_v   2.18   16 Oct 1995 16:47:56   JOE_BOSTIC  $ */
-/***********************************************************************************************
- ***              C O N F I D E N T I A L  ---  W E S T W O O D  S T U D I O S               ***
- ***********************************************************************************************
- *                                                                                             *
- *                 Project Name : Command & Conquer                                            *
- *                                                                                             *
- *                    File Name : CREW.H                                                       *
- *                                                                                             *
- *                   Programmer : Joe L. Bostic                                                *
- *                                                                                             *
- *                   Start Date : April 23, 1994                                               *
- *                                                                                             *
- *                  Last Update : April 23, 1994   [JLB]                                       *
- *                                                                                             *
- *---------------------------------------------------------------------------------------------*
- * Functions:                                                                                  *
- * - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
-
-#ifndef CREW_H
-#define CREW_H
-
-/****************************************************************************
-**	This class handles the basic crew logic. This includes hero tracking,
-**	crew bail-out, and attached object logic.
-*/
-class CrewClass
-{
-public:
-    /*
-    **	This keeps track of the number of "kills" the unit as accumulated.
-    **	When it reaches a certain point, the unit improves.
-    */
-    unsigned short Kills;
-
-    /*
-    **	Constructors, Destructors, and overloaded operators.
-    */
-    CrewClass(void)
-    {
-        Kills = 0;
-    };
-    CrewClass(NoInitClass const&)
-    {
-    }
-
-    int Made_A_Kill(void)
-    {
-        Kills++;
-        return (Kills);
-    };
-
-private:
-};
-
-#endif
