+++ tiberiandawn/function.h
@@ -501,7 +501,7 @@ void Load_Recording_Values(void);
 */
 void* Small_Icon(void const* iconptr, int iconnum);
 void Set_Window(int window, int x, int y, int w, int h);
-long Load_Uncompress(FileClass& file, BuffType& uncomp_buff, BuffType& dest_buff, void* reserved_data);
+long Load_Uncompress(FileClass& file, BufferClass& uncomp_buff, BufferClass& dest_buff, void* reserved_data);
 long Translucent_Table_Size(int count);
 void* Build_Translucent_Table(void const* palette, TLucentType const* control, int count, void* buffer);
 void* Conquer_Build_Translucent_Table(void const* palette, TLucentType const* control, int count, void* buffer);
