+++ tiberiandawn/jshell.cpp
@@ -191,7 +191,7 @@ void File_Fatal(char const* message)
  * HISTORY:                                                                                    *
  *   10/17/1994 JLB : Created.                                                                 *
  *=============================================================================================*/
-long Load_Uncompress(FileClass& file, BuffType& uncomp_buff, BuffType& dest_buff, void* reserved_data)
+long Load_Uncompress(FileClass& file, BufferClass& uncomp_buff, BufferClass& dest_buff, void* reserved_data)
 {
     unsigned short size;
     void* sptr = uncomp_buff.Get_Buffer();
